third_party/blink/web_tests/http/tests/images/document-policy-image-policies-with-border-radius-expected.txt

CONSOLE ERROR: Document policy violation: unoptimized-lossy-images is not allowed in this document.
